app/requirements.txt
fastapi==0.135.2 #Framework web para la API REST
uvicorn==0.24.0 # Servidor ASGI para ejecutar la aplicación FastAPI
sqlalchemy==2.0.36 # ORM para interactuar con la base de datos PostgreSQL
psycopg2-binary==2.9.11 # Adaptador PostgreSQL para Python un driver para SQLAlchemy
python-dotenv==1.0.0 # Carga variables de entorno desde un archivo .env
pydantic==2.12.0 # Validación de datos
pytest==7.4.3 # Framework de pruebas para Python osea testing unitario
pytest-cov==4.1.0 # Plugin de pytest para medir la cobertura de código
pytest-asyncio==1.3.0 # Plugin de pytest para manejar pruebas asincrónicas
boto3==1.29.7 # SDK de AWS para Python, necesario para interactuar con AWS Cognito
PyJWT==2.10.1 # Biblioteca para trabajar con JSON Web Tokens (JWT), utilizada para la autenticación
requests==2.31.0 # Biblioteca para hacer solicitudes HTTP, útil para pruebas y para interactuar con servicios externos como AWS Cognito
httpx==0.25.2 # Cliente HTTP usado por TestClient de FastAPI para testing de endpoints
starlette==1.0.0 # ASGI framework mejorado con mejor compatibilidad
